Added backward selections & mc first bindings
For now b selects to previous word start (which is common vim behavior)
and v selects to previous word end (which didn't realy exist). We'll see
if such a binding is useful.

Added first 3 multiple cursor commands - to mark next like current one
(which just duplicates the cursor if there's no selection yet), to split
a region into lines and to regex select stuff in a region. More to come!