Lexical binding and new next-found-char definition
Moving to lexical binding as it is closer to how my mind works.
The select/extend to next found char function doesn't fit into the
motion macro and I don't see a good way to enhance it in a way that
would make it fit so it's written out by hand for now.